Add function that calls addAdmin() on user creation (#1829)
* Add createUser function that calls addAdmin(user.uid) on user creation

* Move and rename autoAdminOnRegister() cloud function

* Move utility functions to role.js and add functions.logger.log() to autoAdminOnRegister

* Move ./utils/role.js -> ./api/roles.js

* Add email to log output of autoAdminOnRegister()
