+++ functions/.runtimeconfig.json.example
@@ -14,5 +14,8 @@
   "agora": {
     "app_id": "",
     "app_certificate": ""
+  },
+  "flag": {
+    "autoadmin": ""
   }
 }
